requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
#psycopg2-binary==2.9.9
alembic==1.13.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.9
pydantic[email]==2.7.1
pydantic-settings==2.2.1
slowapi==0.1.9
httpx==0.27.0
redis==5.0.4
openai==1.35.0
pgvector==0.3.2
pytest==8.2.0
pytest-asyncio==0.23.6
pytest-cov==5.0.0
aiosqlite==0.20.0